uppsrc: OLD_VALUE completely removed, Core/Rpc: Now supports Json parameters and return values
git-svn-id: svn://ultimatepp.org/upp/trunk@5356 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Draw/DrawTextUtil.cpp
@@ -11,7 +11,7 @@ void DrawTextEllipsis(Draw& w, int x, int y, int cx, const wchar *text, const ch
 	int dtl = 0;
 	int el = 0;
 	for(s = ellipsis; *s; s++) {
-		dtl += f[(byte)*s];
+		dtl += f[*s];
 		el++;
 	}
 	int l = 0;
@@ -20,7 +20,7 @@ void DrawTextEllipsis(Draw& w, int x, int y, int cx, const wchar *text, const ch
 		l += f[(byte) text[i]];
 		if(l > cx) {
 			while(l + dtl > cx && i > 0) {
-				l -= f[(byte) text[i]];
+				l -= f[text[i]];
 				i--;
 			}
 			i++;
